requirements.txt
git+https://github.com/Rapptz/discord.py
pynacl
git+https://github.com/imayhaveborkedit/discord-ext-voice-recv.git
numpy
scipy
pydub
python-dotenv
aiofiles
pillow
requests
google-genai
